requirements.txt
bilibili_api==9.1.0
fire==0.7.0
pydantic==1.10.14
quart==0.19.8
Requests==2.32.3
sqlmodel==0.0.22

nonebot2>=2.2.0
nonebot-adapter-onebot>=2.4.0
nb-cli>=1.0.0
aiohttp>=3.8.0
pillow>=9.0.0
sqlalchemy>=2.0.0
asyncio>=3.4.3
nonebot-plugin-apscheduler>=^0.3.0
